Exchange RLPReader source to the origin project
The so far defined source the RLPReader.sol contract linked to much
larger project which also use the reader. To avoid redirections,
providing the origin source is better.
More benefits are that this source includes a package, which may be used
in the future. Furthermore this project contains tests, the previous
source does not, which can be reused here as orientation to have full
coverage.

Related to: #4
See also: #5, #2, #6